Support startup APs by EFI_MP_SERVICES
Due to UEFI BIOS design change. The APs should be waked up by EFI_MP_SERVICES
which provided by UEFI BIOS.

Interface with Kernelflinger(tos_startup_info_t) must be changed:
	1. new field: system_table_addr
	2. new field: attkb_key[]
	3. TOS_STARTUP_VERSION increased from 2 to 3
	4. use fill TOS_STARTUP_VERSION to tos_image_header to let bootloader
	   identify correct interface version.

Tracked-On: OAM-73461
Reviewed-on: 654952

+++ lib/mp_init/mp_init.c
@@ -196,11 +196,11 @@ uint32_t launch_aps(uint32_t sipi_page, uint8_t expected_cpu_num, uint64_t c_ent
 		return 1;
 
 	/* prepare for sipi */
-	for (i=0; i<=(MAX_CPU_NUM); i++) {
+	for (i=0; i <= (MAX_CPU_NUM); i++) {
 		esp = (uint32_t)(uint64_t)(&ap_stack[i*STARTAP_STACK_SIZE]);
 		setup_cpu_startup_stack(i, esp);
 	}
-	setup_sipi_page(sipi_page, FALSE, (uint64_t)c_entry);
+	setup_sipi_page(sipi_page, FALSE, c_entry);
 
 	/* send sipi */
 	wakeup_aps(sipi_page);
